Extend the clickable area of the file name to include the thumbnail
Now that the checkbox was moved to its own column clicking on the
thumbnail should behave like clicking on the file name. To achieve this
the left position was replaced with a padding, so the element is kept at
the same place while extending its clickable area to cover the
thumbnail.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>