Add a share when clicking on the confirm button
Clicking on the confirm button now adds a share, but only if there is
just a single exact match. If there are no exact matches or there is
more than one exact match no share is added, and the autocomplete
dropdown is shown again with all the suggestions.

The autocomplete dropdown shows sharees already shared with when the
sharee has an e-mail address; in those cases clicking on the recipient
resends the mail notification. In a similar way, when confirming a
single exact match for a sharee already shared with that has an e-mail
address the mail notification is sent again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>